LPD-57577 Move fetch to the root of the component tree
