+++ modules/apps/notification/notification-web/src/main/resources/META-INF/resources/js/components/SettingsContainer/EmailNotificationSettings.tsx
@@ -4,14 +4,16 @@
  */
 
 import ClayPanel from '@clayui/panel';
-import {FormError} from '@liferay/object-js-components-web';
+import {FormError, MultiSelectItem} from '@liferay/object-js-components-web';
 import {ILearnResourceContext} from 'frontend-js-components-web';
-import React from 'react';
+import React, {useEffect, useState} from 'react';
 
 import {NotificationTemplateError} from '../EditNotificationTemplate';
 import {PrimaryRecipient} from './PrimaryRecipients';
 import {SecondaryRecipient} from './SecondaryRecipients';
 import {Sender} from './Sender';
+import {getEmailNotificationRoles} from './rolesUtil';
+import {getUserGroups} from './userGroupsUtil';
 
 import './EmailNotificationSettings.scss';
 
@@ -55,6 +57,22 @@ export function EmailNotificationSettings({
 	setValues,
 	values,
 }: EmailNotificationSettingsProps) {
+	const [roles, setRoles] = useState<MultiSelectItem[]>([]);
+	const [userGroups, setUserGroups] = useState<MultiSelectItem[]>([]);
+
+	useEffect(() => {
+		const fetchInitialData = async () => {
+			const emailNotificationRoles =
+				await getEmailNotificationRoles(baseResourceURL);
+			const emailNotificationUserGroups = await getUserGroups();
+
+			setRoles(emailNotificationRoles);
+			setUserGroups(emailNotificationUserGroups);
+		};
+
+		fetchInitialData();
+	}, [baseResourceURL]);
+
 	return (
 		<div className="lfr__notification-template-email-notification-settings">
 			<ClayPanel
@@ -77,7 +95,6 @@ export function EmailNotificationSettings({
 			>
 				<ClayPanel.Body>
 					<PrimaryRecipient
-						baseResourceURL={baseResourceURL}
 						errors={errors}
 						learnResources={learnResources}
 						recipientOptions={
@@ -85,8 +102,10 @@ export function EmailNotificationSettings({
 								? [...RECIPIENT_OPTIONS, SUBSCRIBERS_OPTION]
 								: RECIPIENT_OPTIONS
 						}
+						roles={roles}
 						selectedLocale={selectedLocale}
 						setValues={setValues}
+						userGroups={userGroups}
 						values={values}
 					/>
 				</ClayPanel.Body>
@@ -98,11 +117,12 @@ export function EmailNotificationSettings({
 			>
 				<ClayPanel.Body>
 					<SecondaryRecipient
-						baseResourceURL={baseResourceURL}
 						learnResources={learnResources}
 						recipientOptions={RECIPIENT_OPTIONS}
+						roles={roles}
 						selectedLocale={selectedLocale}
 						setValues={setValues}
+						userGroups={userGroups}
 						values={values}
 					/>
 				</ClayPanel.Body>
